refactor: split cli/prompt into cached, refresh, direct submodules
Separate the four responsibilities in prompt.rs into dedicated files:
- cached.rs: run_cached + spawn_refresh_if_needed (stale-while-revalidate)
- refresh.rs: run_refresh (background cache worker)
- direct.rs: run_direct (no-cache direct execution)
prompt.rs retains only run() dispatch
